feat: add WorkspaceTabView component and related hooks for improved tab management
- Implemented WorkspaceTabView to handle different tab types: auth-users, js-query, and firestore-query.
- Created useAppCommandPaletteController hook to manage command palette interactions.
- Added tests for useAppShellController to ensure proper tab management and interactions.
- Developed useAppShellController hook to manage application shell state, including tabs, projects, and activity logging.
